Normalize to integer instead of float

+++ src/probability_normalizer.rb
@@ -1,7 +1,7 @@
 class ProbabilityNormalizer
   
   def normalize(weight)
-    weight == 0.0 ? -9.223372036854776E18 : -(Math.log(weight)*10).round.to_f 
+    weight == 0.0 ? -9.223372036854776E18 : -(Math.log(weight)*10).round.to_i
   end
   
 end
